Added LinkTextSpan
